api/requirements.txt
aniso8601==8.0.0
azure-core==1.1.1
azure-cosmos==4.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
idna==3.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.23.5
pytz==2020.1
requests==2.28.1
six==1.15.0
urllib3==1.26.12
Werkzeug==1.0.1
